src/surquest/GCP/requirements.txt
# google-auth==2.40.3
# google-auth-oauthlib==1.2.2
# google-api-python-client==2.173.0
google-cloud-dataform==0.6.2
pathspec==0.12.1
